Refactor: 필터 섹션과 버튼 컴포넌트를 common으로 이동
