+++ src/components/common/index.ts
@@ -1,3 +1,5 @@
 export { default as Header } from './Header'
+export { default as Button } from './Button'
+export { default as FilterSection } from './FilterSection'
 export { default as Avatar } from './Avatar'
 export { styles } from './common.styles'
