+++ src/components/teamPage/index.ts
@@ -1,5 +1,4 @@
-export { default as Button } from './Button'
 export { default as MemberList } from './MemberList'
 export { default as TeamContent } from './TeamContent'
-export { default as FilterSection } from './FilterSection'
+export { default as MemberCard } from './MemberCard'
 export { styles } from './TeamPage.styles'
